% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviateSpeciesName.R
\name{abbreviate_species}
\alias{abbreviate_species}
\title{Converts long biological names to their short version}
\usage{
abbreviate_species(names)
}
\arguments{
\item{names}{a character vector to be converted}
}
\value{
a character vector
}
\description{
This function converts long botanical names to their short variant.
}
