% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Transform first letter of a word or words to uppercase}
\usage{
capitalize(char_vec, all.fragments = TRUE)
}
\arguments{
\item{char_vec}{a character vector to be converted}

\item{all.fragments}{a logical value indicating if all the word fragments should be converted. It default to TRUE}
}
\value{
a character vector
}
\description{
This function converts the first letter on a word to uppercase.
}
