% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractChars.R
\name{extract_chars}
\alias{extract_chars}
\title{Extract characters of interest}
\usage{
extract_chars(chrs_vec, regx = "[a-zA-Z]+", bind_with = ".")
}
\arguments{
\item{chrs_vec}{a character vector}

\item{regx}{a regex pattern}

\item{bind_with}{an optional character that will be used to bind the fragments}
}
\value{
a character vector
}
\description{
This function extracts characters matching the regex pattern.
It is generally a simple wrapper around the stringi::extract_all
package.
}
