% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripWhiteSpaces.R
\name{strip_white_spaces}
\alias{strip_white_spaces}
\title{Strip white spaces from characters}
\usage{
strip_white_spaces(chr_vec, replace_with = "")
}
\arguments{
\item{chr_vec}{a character vector}

\item{replace_with}{the character that will be used to replace the white spaces}
}
\value{
a character vector
}
\description{
This function removes white spaces from characters contained in a vector
}
